Fix many verbose compiler warnings
Many compiler warnings regarding implicit conversions.

+++ r5dev/engine/client/cl_rcon.cpp
@@ -127,7 +127,7 @@ void CRConClient::Disconnect(void)
 void CRConClient::Send(const string& svMessage) const
 {
 	ostringstream ssSendBuf;
-	const u_long nLen = htonl(svMessage.size());
+	const u_long nLen = htonl(static_cast<u_long>(svMessage.size()));
 
 	ssSendBuf.write(reinterpret_cast<const char*>(&nLen), sizeof(u_long));
 	ssSendBuf.write(svMessage.data(), svMessage.size());
